devtools: extract network related into new file
